Funcional todos los cruds de facturas. continuo con la vista para imprimir

+++ variedadeselpaisa/database/migrations/2021_05_13_215318_create_factura_table.php
@@ -17,7 +17,7 @@ class CreateFacturaTable extends Migration
             $table->bigIncrements('id');
             $table->integer('usuario_id')->unsigned();
             $table->bigInteger('tfra_id')->unsigned();
-            $table->integer('valor');
+            $table->bigInteger('valor');
             $table->text('concepto')->nullable();
             $table->tinyInteger('estado');
             $table->timestamps();
